Add util function writing installed extension to db
