+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './logger';
 export * from './events';
 export * from './stats';
+export * from './extensions';
